+++ src/app/(aboutme)/interests/page.tsx
@@ -2,13 +2,16 @@
  * @author Adam Long
  * @copyright 2024 - All rights reserved.
  */
+
+//Imports
 import {BackgroundNonHome} from "@/components/backgroundNonHome";
 import {NavbarNonHome} from "@/components/navbarNonHome";
 import GradualSpacing from "@/components/ui/gradualSpacing";
-import {philosophyElements} from "@/app/philosophies/philosophyElements";
-import PhilosophyCard from "@/components/philosophyCard";
+import InterestCard from "@/components/interestCard";
+import {interestElements} from "@/app/(aboutme)/interests/interestElements";
 
-function Philosophies() {
+function Interests() {
+
     return (
         <div className="relative min-h-screen min-w-screen overflow-hidden">
             {/*Background component*/}
@@ -21,39 +24,31 @@ function Philosophies() {
                 <div className='grid grid-cols-3 pt-20'>
                     <div className='flex flex-col items-start'>
                         <GradualSpacing
-                            text="PHILOSOPHIES"
+                            text="INTERESTS"
                             className="font-display text-sm text-white font-avalors shadow-inner"
                             duration={2.0}
                         />
                         <br/>
-                        <div className='bg-gray-900 h-1 w-full'/>
+                        <div className='bg-gray-900 h-1 w-full'></div>
                         <br/>
                         <GradualSpacing
-                            text="How I live my life"
+                            text="That's what I like"
                             className="font-display text-3xl text-white font-avalors shadow-inner"
                             duration={2.0}
                         />
                     </div>
                 </div>
                 {/*Page Content*/}
-                <div className="relative flex flex-col items-center justify-center w-9/12 mt-32">
-                    <h1 className='text-2xl shadow-inner text-center font-aptos text-gray-300 leading-relaxed'>
-                        My life philosophies are guided by lessons I have learned. They are written on my mirror, so
-                        that I can be a reflection of who I want to be.
-                    </h1>
-                    <div className='bg-gray-900 h-1 w-full m-8'/>
-                    {/*Row-By-Row Cards*/}
-                    <div className='mt-10 w-full'>
-                        {philosophyElements.map((elements) => (
-                            <div key={elements.title} className='p-4'>
-                                <PhilosophyCard prop={elements}/>
-                            </div>
-                        ))}
-                    </div>
+                <div className="w-10/12 px-8 py-20 space-y-32 mt-24">
+                    {interestElements.map((interest) => (
+                        <div key={interest.title}>
+                            <InterestCard props={interest}/>
+                        </div>
+                    ))}
                 </div>
             </div>
         </div>
     );
 }
 
-export default Philosophies;
+export default Interests;
