+++ src/app/(aboutme)/philosophies/philosophyElements.ts
