+++ src/app/(aboutme)/whoAmI/textElements.ts
